🎲

+++ cmd/cron/delete_ipa/delete_interim_ipa.go
@@ -2,10 +2,14 @@ package delete_ipa
 
 import (
 	"context"
+	"dumpapp_server/pkg/common/clients"
+	"dumpapp_server/pkg/common/constant"
 	"dumpapp_server/pkg/common/util"
+	"dumpapp_server/pkg/config"
 	impl2 "dumpapp_server/pkg/controller/impl"
 	"dumpapp_server/pkg/dao/impl"
 	"dumpapp_server/pkg/dao/models"
+	util2 "dumpapp_server/pkg/util"
 	"fmt"
 	"github.com/volatiletech/sqlboiler/v4/queries/qm"
 	"time"
@@ -14,6 +18,12 @@ import (
 func Run() {
 	fmt.Println("DeleteInterimIpa")
 	run()
+}
+
+type deleteIpa struct {
+	ID      int64
+	Name    string
+	Version string
 }
 
 func run() {
@@ -25,6 +35,12 @@ func run() {
 
 	tm := time.Now().AddDate(0, 0, -3)
 
+	txn := clients.GetMySQLTransaction(ctx, clients.MySQLConnectionsPool, true)
+	defer clients.MustClearMySQLTransaction(ctx, txn)
+	ctx = context.WithValue(ctx, constant.TransactionKeyTxn, txn)
+
+	deleteIpaMap := make(map[int64][]*deleteIpa, 0)
+
 	for hasNext {
 		fmt.Println(fmt.Sprintf("offset: %d...", offset))
 
@@ -35,22 +51,54 @@ func run() {
 		ids, err := impl.DefaultIpaDAO.ListIDs(ctx, offset, bulkSize, filters, nil)
 		util.PanicIf(err)
 
+		ipaMap, err := impl.DefaultIpaDAO.BatchGet(ctx, ids)
+		util.PanicIf(err)
+
 		hasNext = len(ids) >= bulkSize
 		offset += len(ids)
 
-		fmt.Println(ids)
-
-		for _, id := range ids {
-			util.PanicIf(impl.DefaultIpaDAO.Delete(ctx, id))
+		for _, ipa := range ipaMap {
+			util.PanicIf(impl.DefaultIpaDAO.Delete(ctx, ipa.ID))
 		}
 		ipaVersions, err := impl.DefaultIpaVersionDAO.BatchGetIpaVersions(ctx, ids)
 		util.PanicIf(err)
 
-		for _, versions := range ipaVersions {
+		for ipaID, versions := range ipaVersions {
+			ipa, ok := ipaMap[ipaID]
+			if !ok {
+				continue
+			}
 			for _, version := range versions {
 				util.PanicIf(impl.DefaultIpaVersionDAO.Delete(ctx, version.ID))
 				util.PanicIf(impl2.DefaultTencentController.DeleteFile(ctx, version.TokenPath))
+				deleteIpaMap[ipaID] = append(deleteIpaMap[ipaID], &deleteIpa{
+					ID:      ipaID,
+					Name:    ipa.Name,
+					Version: version.Version,
+				})
 			}
 		}
 	}
+
+	clients.MustCommit(ctx, txn)
+	util.ResetCtxKey(ctx, constant.TransactionKeyTxn)
+
+	nameStr := ""
+	for _, is := range deleteIpaMap {
+		for _, ipa := range is {
+			nameStr += fmt.Sprintf("<font color=\"info\">名称：</font>%s %s\n>", ipa.Name, ipa.Version)
+		}
+	}
+
+	contentStr := fmt.Sprintf("<font color=\"warning\">已删除临时 ipa：</font>\n>")
+	deleteIpaStr := fmt.Sprintf("<font color=\"comment\">%s</font>", nameStr)
+	data := map[string]interface{}{
+		"msgtype": "markdown",
+		"markdown": map[string]interface{}{
+			"content": contentStr +
+				deleteIpaStr,
+		},
+	}
+	util2.SendWeiXinBot(ctx, config.DumpConfig.AppConfig.TencentGroupKey, data, []string{"@all"})
+
 }
